Authentication and Authorization in Postman

+++ controllers/api/v1/posts_api.js
@@ -29,20 +29,21 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
 
-        post.remove();
-        
-        await Comment.deleteMany({post : req.params.id});
+        if(post.user == req.user.id){
+            post.remove();
 
-        console.log('Post and comments deleted.');
-        return res.status(200).json({
-            message : "Successfully deleted the post.",
-        });
+            await Comment.deleteMany({post : req.params.id});
+
+            console.log('Post and comments deleted.');
+            return res.status(200).json({
+                message : "Successfully deleted the post.",
+            });
+        }
         
     }catch(error){
-
         console.log('Error : ', error);
-        return res.status(500).json({
-            message : "Post not be deleted."
+        return res.status(401).json({
+            message : "You cannot delete this post."
         });
     }
 }
